Fix seeking behaviour to find the exact frame.
When seeking, the source gives a keyframe which might be a couple of
seconds before the requested timestamp. This commit allows the audio
drivers to discard all the audio packet they receive from the source
until they reach the correct timestamp.

At the moment DirectSound, OpenAL, and the SilentAudioPlayer are
implemented. OpenAL could not be tested.

Still need to implement Pulse audio driver seeking method.