Fix video texture anchor point
The Player need to flip the video texture along the y axis. But
it was surprising to find the anchor point flipped to the top.
Now the image is flipped but the anchor point remains at the
bottom left of the texture, which is the normal behaviour for
OpenGL textures.