Fix bug introduced in PR #104 (commit af034f2)
If set_exclusive_mouse is called from within on_mouse_press event handler,
the mouse release event is not called which results in not calling
_user32.ReleaseCapture().

This means that the mouse remained capture by the window, and created weird
behaviour if the user tried to click on another window or on the window title
bar.

Now _user32.ReleaseCapture() is called when set_exclusive_mode is called to
make sure it's released.