EventDispatcher and Clock only hold weakref to methods
Both EventDispatcher and pyglet Clock held a reference to any method provided as
argument. This meant that a reference to the instance was created which did not
allow the object to be garbage collected immediately, as the user did not know
that an extra reference was still keeping the object alive.

Now the EventDispatcher and the Clock only hold a WeakMethod reference to the
instance. When the method is called, we check if the object is still alive. In
case it's not, we remove the method from the list of handlers (EventDispatcher)
or from the scheduled functions (Clock).