Fix event handler not throwing all TypeError exceptions
In Dipatcher class, `_raise_dispatch_exception` is used to provide better error
messages. Problem was that if the TypeError was not due to a wrong number of
arguments, the exception was silently swallowed. This created nasty behaviour,
especially as `on_draw` method calls many OpenGL functions which could throw a
TypeError when the arguments are not correctly used.