Fix infinite recusion during window operations on macOS
cocoapy function `send_super()` which is supposed to call a method on the superclass sometimes calls the method on the object class instead leading to infinite recursion. I have no idea why this happens, but I isolated to events on which it happens, and forced the function to call grand-parent instead of the parent class, which seems to resolve the issue.

In the long term this should probably be solved either by someone figuring out how cocoapy works, or by switching to PyObjC which seem better supported.

Bug #5