Implement @event.intercept annotation for higher-priority event handlers.
All handlers with this annotation are invoked before the handlers without this annotation. Unlike normal handlers, intercept handlers are invoked in order of registration, i.e. the earlier the handler is registered, the higher priority it has.

Also when an unknown event is dispatched, EventException is raised instead of AssertionError.