Atlas border fix and enhancement (#216)
* Fix issue where TextureBin was not passing border attribute to TextureAtlas.
Added atlas_border attribute to the resource.image function if you wish to specify if it adds pixel borders.
Made change where border attribute is the amount of pixels to separate, rather than enabled as true or false. This allows more pixels to be padded in rare cases where 1 pixel is not enough of a border. 0 can still be set for no border.

* Allow animations to specify borders as well.