Added a new `shapes.Star` shape
This new shape has a configurabel number of spikes, as well as
an inner and outer radius. That way, not just stars can be rendered,
but also theoretically a rhombus (with num_spikes=2).

The code is kept in a very similar fashion to the
existing shapes.

The change also include a few minor typo fixes in the rest of the file.

An example for a star is added to the module docstring.