Fix SpriteGroup batching by computing correct hash (#706)
While SpriteGroup __eq__ implementation no longer relies on the id of a parent, the __hash__ does. This still breaks grouping of multiple SpriteGroups when the parents are different instances (different id).
The reason is that Batch object checks for inclusion into maps/sets (Batch.group_map, Batch.group_children ...), which checks for both the hash and the equality of the items.