Add method to encapsulate shapes vertex creation
ShapeBase._create_vertex_list() is an abstract method that should be overridden to provide the creation code for the shape's vertex list. All existing shapes have been adapted to this behaviour.

Extraction of the vertex list creation code is necessary to later obtain group and batch writable properties, as migration of the group sometimes requires vertex list destruction and re-creation.

An additional class attribute _draw_mode is added, and is required so that the shape can retrieve its draw mode every time it is needed (e.g. during vertex list creation and draw).