3.12 fix backport for 1.5 (#979)
Use correct data types
Define BYTE using the correct type when using an older Python version.
Fix expected data type of BYTE for IWICBitmapFrameEncode data.
Add a warning if the GDIFont cannot find a font and reverts to default.
Add a catch and check for a localized font name from GDI enumeration. This fixes an issue with the font_comparison where fonts with @ in front were falling back to the default font.