Fixes for Python 3.12 (#977)
* Use correct data type. Fixes: https://github.com/pyglet/pyglet/issues/964

* Define BYTE using the correct type when using an older Python version.
Fix expected data type of BYTE for IWICBitmapFrameEncode data.
Fix data type of XINPUT_GAMEPAD to the correct label.

* Add a warning if the GDIFont cannot find a font and reverts to default.
Add a catch and check for a localized font name from GDI enumeration. This fixes an issue with the font_comparison where fonts with @ in front were falling back to the default font.