Initial change of shapes.Arc from GL_LINES to computed geometry.
The Arc class was drawn using GL_LINES as the drawing mode.  This
ment that if you wanted to change the width of the arc you had
to call glLineWidth which would result in artifacts at higher
widths.  Now the Arc is constructed from quads in a similar fassion
as the Line class.

There is still potential optimizations to be made and the "miter"
lines are not always scaled properly in this version.