Arc Shape improvements (#1028)
* Initial change of shapes.Arc from GL_LINES to computed geometry.

The Arc class was drawn using GL_LINES as the drawing mode.  This
ment that if you wanted to change the width of the arc you had
to call glLineWidth which would result in artifacts at higher
widths.  Now the Arc is constructed from quads in a similar fassion
as the Line class.

There is still potential optimizations to be made and the "miter"
lines are not always scaled properly in this version.

* Fixed scaling issue for the miter vectors of differing angles

* Moved the width argument to after closed

* Cleaned up corner cases and improved performance for Arcs

- Improved performance of the Arc generation by about 50% from the
  first pass implementation.
- Handled more corner cases related to closed and not closed
  Arcs

* Removed some debug code

* Made changes realted to comments in pull request.

- Changed width to thickness
- Added spaces to return values of _get_segment

---------

Co-authored-by: Kate Winsor <katewinsor@Kates-MacBook-Pro.local>
Co-authored-by: Adam Winsor <adamwinsor@Kates-MacBook-Pro.local>